# swagger.yml
openapi: 3.0.3
info:
  title: Event Management API
  description: API for managing events, editions, and activities
  version: 1.0.0
security:
  - basicAuth: []
servers:
  - description: Example server
    url: https://example.com/api/v1
...
paths:
  /events:
    get:
      tags:
        - admin
        - organizer
        - user
      summary: View all registered events
      operationId: listEvents
      responses:
        "200":
          description: Registered events showed succesfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
    post:
      tags:
        - admin
      summary: Register an event
      operationId: registerEvent
      requestBody:
        description: Event data to be registered
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventWithoutId"
      responses:
        "201":
          description: Event registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        "400":
          description: Invalid data provided
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
...
  /events/{eventId}/editions:
    get:
      tags:
        - admin
        - organizer
        - user
      summary: View all editions of an event
      operationId: viewEditions
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event to be viewed
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Registered event showed succesfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Edition'
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
        "404":
          description: Event not found
    post:
      tags:
        - admin
      summary: Register an edition for an event
      operationId: registerEdition
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event to be used
          schema:
            type: integer
            format: int64
      requestBody:
        description: Edition data to be registered
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditionWithoutId"
      responses:
        "201":
          description: Edition registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edition'
        "400":
          description: Invalid data provided
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
        "404":
          description: Event not found
  /events/{eventId}/editions/{editionId}:
    get:
      tags:
        - admin
        - organizer
        - user
      summary: View an edition of an event
      operationId: viewEdition
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event
          schema:
            type: integer
            format: int64
        - name: editionId
          in: path
          required: true
          description: ID of the edition to be viewed
          schema:
            type: string
      responses:
        "200":
          description: Event edition showed succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edition"
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
        "404":
          description: Event edition not found
    delete:
      tags:
        - admin
      summary: Remove an edition of an event
      operationId: removeEdition
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event
          schema:
            type: integer
            format: int64
        - name: editionId
          in: path
          required: true
          description: ID of the edition to be deleted
          schema:
            type: string
      responses:
        "204":
          description: Event edition deleted successfully
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
        "404":
          description: Event edition not found
    put:
      tags:
        - admin
        - organizer
      summary: Edit an edition of an event
      operationId: editEdition
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event to be edited
          schema:
            type: integer
            format: int64
        - name: editionId
          in: path
          required: true
          description: ID of the edition to be edited
          schema:
            type: string
      requestBody:
        description: Updated edition data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditionWithoutId"
      responses:
        "200":
          description: Event edition edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edition'
        "400":
          description: Invalid data provided
        "401":
          description: User must pass valid credentials to perform this action
        "403":
          description: User not authorized to perform this action
        "404":
          description: Event edition not found
...
components:
  schemas:
    EventWithoutId:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        description:
          type: string
      required:
        - name
        - abbreviation
        - description
    Event:
      allOf:
        - $ref: '#/components/schemas/EventWithoutId'
      type: object
      properties:
        id:
          type: integer
          format: int64
      required:
        - id
    EditionWithoutId:
      type: object
      properties:
        number:
          type: integer
          format: int64
        year:
          type: integer
          format: int64
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        city:
          type: string
        organizer_id:
          type: string # Organizers won't be able to edit this.
      required:
        - number
        - year
        - start_date
        - end_date
        - city
        - organizer
    Edition:
      allOf:
        - $ref: '#/components/schemas/EditionWithoutId'
      type: object
      properties:
        id:
          type: integer
          format: int64
      required:
        - id
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic